+++ backend/routes/auth.routes.js
@@ -5,4 +5,6 @@ const authServices = require("../services/auth.services.js");
 // Route de connexion employé
 router.post("/login", authServices.loginEmployee);
 
+router.post("/create-account", authServices.createAccount);
+
 module.exports = router;